+++ mobile/android/android-components/components/feature/awesomebar/src/main/java/mozilla/components/feature/awesomebar/provider/HistoryStorageSuggestionProvider.kt
@@ -11,6 +11,7 @@ import mozilla.components.browser.icons.BrowserIcons
 import mozilla.components.browser.icons.IconRequest
 import mozilla.components.concept.awesomebar.AwesomeBar
 import mozilla.components.concept.engine.Engine
+import mozilla.components.concept.engine.EngineSession.LoadUrlFlags
 import mozilla.components.concept.storage.HistoryStorage
 import mozilla.components.concept.storage.SearchResult
 import mozilla.components.feature.awesomebar.facts.emitHistorySuggestionClickedFact
@@ -150,7 +151,10 @@ internal suspend fun Iterable<SearchResult>.into(
             editSuggestion = if (showEditSuggestion) result.url else null,
             score = result.score,
             onSuggestionClicked = {
-                loadUrlUseCase.invoke(result.url)
+                loadUrlUseCase.invoke(
+                    result.url,
+                    flags = LoadUrlFlags.select(LoadUrlFlags.LOAD_FLAGS_DISALLOW_INHERIT_PRINCIPAL),
+                )
                 emitHistorySuggestionClickedFact()
             },
         )